# wms-contract/src/main/resources/openapi/wms-v1.yaml
openapi: 3.1.0
info:
  title: WMS v1 HTTP contract
  version: 0.1.0
  description: |
    公开前缀 `/api/wms/v1`，内部前缀 `/internal/wms/v1`。
    enterpriseId/operator 来自已验证 OIDC token；请求中的企业字段仅核验一致性，不能扩大权限。
    数量 JSON 为十进制字符串。时间 RFC3339 UTC。列表 cursor 分页，默认 50、上限 200。
    写入头必须带 Authorization、Idempotency-Key、X-Request-Id；clientOperationId 若出现必须与 Idempotency-Key 一致。
    Seata TCC Try/Confirm/Cancel 不是 REST 资源，不提供自定义 /prepare 或 /decisions。
    OIDC issuer/client/权限映射由环境变量配置，本文件不含密钥。
    本契约覆盖 S1 主数据与后续切片接口形状；未实现切片不得把 200 当成业务已交付。
servers:
  - url: /
security:
  - oidc: []
...
paths:
...
  /api/wms/v1/warehouses/{warehouseId}/count-plans:
    post:
      operationId: createCountPlan
      tags: [inventory]
      summary: 创建盘点计划
      security:
        - oidc: [count.create]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountPlanCreateRequest'
      responses:
        '201':
          description: 已创建资源
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
    get:
      operationId: listCountPlans
      tags: [inventory]
      summary: 盘点计划列表
      security:
        - oidc: [count.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursorPage'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /api/wms/v1/warehouses/{warehouseId}/count-plans/{countPlanId}/freeze-requests:
    post:
      operationId: requestCountFreeze
      tags: [inventory]
      summary: 进入QUIESCING
      security:
        - oidc: [count.freeze]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/CountPlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionedReasonRequest'
      responses:
        '202':
          description: 已持久化受理，不是业务终态成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedOperation'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /api/wms/v1/warehouses/{warehouseId}/count-plans/{countPlanId}/observations:
    post:
      operationId: recordCountObservation
      tags: [inventory]
      summary: 盘点观察，不覆盖历史轮次
      security:
        - oidc: [count.record]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
        - $ref: '#/components/parameters/CountPlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountObservationRequest'
      responses:
        '201':
          description: 已创建资源
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /api/wms/v1/warehouses/{warehouseId}:
    get:
      operationId: getWarehouse
      tags: [masterdata]
      summary: 仓库详情
      security:
        - oidc: [masterdata.read]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/WarehouseId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceEnvelope'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      description: issuer 与 client 由 WMS_OIDC_ISSUER / WMS_OIDC_CLIENT_ID 指向已有提供方，仓库不写死密钥
      openIdConnectUrl: https://issuer.example.invalid/.well-known/openid-configuration
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 64 }
      description: 与 clientOperationId 一致；同键同内容返回原结果，同键异内容 409
    XRequestId:
      name: X-Request-Id
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
...
    Cursor:
      description: 不透明游标，绑定原查询条件；更换筛选或投影世代后从第一页重新读取。
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    WarehouseId:
      name: warehouseId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
    CountPlanId:
      name: countPlanId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Id' }
...
  responses:
...
    Error:
      description: 业务或协议错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Id:
      type: string
      minLength: 1
      maxLength: 64
...
    Version:
      type: integer
      format: int64
      minimum: 0
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, requestId, retryable]
      properties:
        code:
          type: string
          enum:
            - INVALID_QUANTITY
            - INVALID_UNIT
            - REQUIRED_LOT
            - INVALID_EXPIRY
            - UNAUTHENTICATED
            - WAREHOUSE_FORBIDDEN
            - RESOURCE_NOT_FOUND
            - IDEMPOTENCY_PAYLOAD_MISMATCH
            - VERSION_CONFLICT
            - STOCK_INSUFFICIENT
            - STOCK_FROZEN
            - EXPIRED_STOCK
            - SERIAL_OWNERSHIP_CONFLICT
            - STALE_OWNER_EPOCH
            - TCC_TRANSACTION_PENDING
            - INVALID_STATE
            - TCC_CONTEXT_MISMATCH
            - TCC_OWNER_CONFLICT
            - ROUTE_EPOCH_STALE
            - RESOURCE_BUDGET_EXCEEDED
            - DEPENDENCY_UNAVAILABLE
            - EFFECT_IDENTITY_CONFLICT
            - STALE_EXECUTION_ATTEMPT
            - EFFECT_ALREADY_APPLIED
            - RECOVERY_PENDING
            - INVALID_ARGUMENT
            - INVALID_PAGE
            - DATABASE_UNAVAILABLE
            - QUERY_OVERLOADED
            - RATE_LIMITED
            - DUPLICATE_DOCUMENT
            - DUPLICATE_INSPECTION
        message: { type: string }
        requestId: { type: string }
        operationId: { $ref: '#/components/schemas/Id' }
        retryable: { type: boolean }
        details:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              field: { type: string }
              code: { type: string }
    CursorPage:
      type: object
      additionalProperties: false
      required: [items, limit]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ResourceEnvelope' }
        nextCursor: { type: string }
        limit: { type: integer, minimum: 1, maximum: 200 }
        asOf: { type: string, format: date-time }
        lagSeconds: { type: integer, minimum: 0 }
    ResourceEnvelope:
      type: object
      additionalProperties: true
      required: [id, version]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        version: { $ref: '#/components/schemas/Version' }
        status: { type: string }
        actions:
          type: array
          items: { type: string }
          description: 交互提示，写入仍以服务端校验为准
    AcceptedOperation:
      type: object
      additionalProperties: false
      required: [operationId, status, statusUrl]
      properties:
        operationId: { $ref: '#/components/schemas/Id' }
        sourceOperationId: { $ref: '#/components/schemas/Id' }
        commandId: { $ref: '#/components/schemas/Id' }
        fulfillmentId: { $ref: '#/components/schemas/Id' }
        receiptId: { $ref: '#/components/schemas/Id' }
        executionId: { $ref: '#/components/schemas/Id' }
        physicalStatus: { type: string }
        stockSyncStatus: { type: string }
        status: { type: string }
        version: { $ref: '#/components/schemas/Version' }
        statusUrl: { type: string }
        safeToRetry: { type: boolean }
...
    VersionedReasonRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        reason: { type: string }
        expectedVersion: { $ref: '#/components/schemas/Version' }
        clientOperationId: { $ref: '#/components/schemas/Id' }
...
    CountPlanCreateRequest:
      type: "object"
      additionalProperties: false
      required: ["locationIds"]
      properties:
        planId:
          type: "string"
          maxLength: 64
        countPlanId:
          type: "string"
          maxLength: 64
        reason:
          type: "string"
          maxLength: 32
        locationIds:
          type: "array"
          items:
            type: "string"
            minLength: 1
            maxLength: 64
          minItems: 1
          maxItems: 200
    CountObservationRequest:
      type: "object"
      additionalProperties: false
      required: ["lineId","qty"]
      properties:
        lineId:
          type: "string"
          maxLength: 64
          minLength: 1
        observationId:
          type: "string"
          maxLength: 64
        qty:
          type: "string"
          pattern: "^[0-9]{1,14}([.][0-9]{1,6})?$"
          description: "精确十进制；不得小于0"
        roundNo:
          type: "integer"
          minimum: 1
...